+++ src/routes/LayoutTestWrapper.svelte
@@ -1,7 +1,0 @@
-<script>
-	import Layout from './+layout.svelte';
-</script>
-
-<Layout>
-	<div data-testid="test-child">Test Content</div>
-</Layout>